# store/cachekv: reduce allocation when convert []byte -> string in map key