@odeke-em 2021-05-06
@cuonglm technically at some point we shall need to file a Go compiler bug to figure out why that performance is not the same given that that idiom should produce the same results.