@cuonglm 2021-05-06
@odeke-em The optimization is only applied if the map index is not on LHS of an assignment, see: https://github.com/golang/go/blob/d2fd503f687ca686cb8fbee0b29e64ba529038fe/src/cmd/compile/internal/walk/order.go#L1173